+++ src/main/java/com/example/bak/community/presentation/CommunityController.java
@@ -1,8 +1,9 @@
 package com.example.bak.community.presentation;
 
-import com.example.bak.community.application.CommunityService;
-import com.example.bak.community.application.dto.CommunityResult;
-import com.example.bak.community.application.dto.CommunityResult.CommunityId;
+import com.example.bak.community.application.command.CommunityCommandService;
+import com.example.bak.community.application.query.CommunityQueryService;
+import com.example.bak.community.application.query.dto.CommunityResult;
+import com.example.bak.community.application.query.dto.CommunityResult.CommunityId;
 import com.example.bak.community.presentation.dto.CommunityRequest;
 import com.example.bak.global.common.response.ApiResponse;
 import com.example.bak.global.common.response.ApiResponseFactory;
@@ -24,13 +25,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1")
 public class CommunityController {
 
-    private final CommunityService communityService;
+    private final CommunityCommandService communityCommandService;
+    private final CommunityQueryService communityQueryService;
 
     @GetMapping("/companies/{companyId}/communities")
     public ResponseEntity<ApiResponse> getCommunities(
             @PathVariable Long companyId
     ) {
-        List<CommunityResult.Detail> communityResults = communityService.getCommunities(companyId);
+        List<CommunityResult.Detail> communityResults = communityQueryService.getCommunities(companyId);
         ApiResponse response = ApiResponseFactory.success("커뮤니티 목록을 성공적으로 불러왔습니다.",
                 communityResults);
         return ResponseEntity.ok(response);
@@ -41,7 +43,7 @@ public class CommunityController {
             @PathVariable Long companyId,
             @RequestBody CommunityRequest request
     ) {
-        CommunityId communityId = communityService.createCommunity(companyId, request.name(),
+        CommunityId communityId = communityCommandService.createCommunity(companyId, request.name(),
                 request.jobGroup());
         ApiResponse response = ApiResponseFactory.successVoid("커뮤니티 정보를 성공적으로 저장했습니다.");
         return ResponseEntity.created(UriUtils.current(communityId.value()))
@@ -53,7 +55,7 @@ public class CommunityController {
             @PathVariable Long communityId,
             @RequestBody CommunityRequest request
     ) {
-        communityService.updateCommunity(communityId, request.name(), request.jobGroup());
+        communityCommandService.updateCommunity(communityId, request.name(), request.jobGroup());
         ApiResponse response = ApiResponseFactory.successVoid("커뮤니티 정보를 성공적으로 수정했습니다.");
         return ResponseEntity.ok(response);
     }
@@ -62,7 +64,7 @@ public class CommunityController {
     public ResponseEntity<ApiResponse> deleteCommunityCompany(
             @PathVariable Long communityId
     ) {
-        communityService.deleteCommunity(communityId);
+        communityCommandService.deleteCommunity(communityId);
         ApiResponse response = ApiResponseFactory.successVoid("회사 정보를 성공적으로 삭제했습니다.");
         return ResponseEntity.ok(response);
     }
